release: Chromium 6.0.472 (6.0.472.9)
Compare: https://github.com/chromium/chromium/compare/2544c4012ca37e2672b0fb656838d1fb0d76a6d4...50544bfe936450d80378b9607a41cb5085046338

+++ chrome/browser/cookies_tree_model.cc
@@ -18,8 +18,6 @@
 #include "base/string_util.h"
 #include "chrome/browser/extensions/extensions_service.h"
 #include "chrome/browser/in_process_webkit/webkit_context.h"
-#include "chrome/browser/net/chrome_url_request_context.h"
-#include "chrome/browser/profile.h"
 #include "grit/app_resources.h"
 #include "grit/generated_resources.h"
 #include "grit/theme_resources.h"
@@ -57,16 +55,14 @@ CookieTreeCookieNode::CookieTreeCookieNode(
 
 void CookieTreeCookieNode::DeleteStoredObjects() {
   // notify CookieMonster that we should delete this cookie
-  // Since we are running on the UI thread don't call GetURLRequestContext().
-  net::CookieMonster* monster = GetModel()->profile_->
-      GetRequestContext()->GetCookieStore()->GetCookieMonster();
   // We have stored a copy of all the cookies in the model, and our model is
   // never re-calculated. Thus, we just need to delete the nodes from our
   // model, and tell CookieMonster to delete the cookies. We can keep the
   // vector storing the cookies in-tact and not delete from there (that would
   // invalidate our pointers), and the fact that it contains semi out-of-date
   // data is not problematic as we don't re-build the model based on that.
-  monster->DeleteCookie(cookie_->first, cookie_->second, true);
+  GetModel()->cookie_monster_->
+      DeleteCookie(cookie_->first, cookie_->second, true);
 }
 
 namespace {
@@ -178,28 +174,25 @@ void CookieTreeLocalStorageNode::DeleteStoredObjects() {
 
 ///////////////////////////////////////////////////////////////////////////////
 // CookieTreeRootNode, public:
-CookieTreeOriginNode* CookieTreeRootNode::GetOrCreateOriginNode(
-    const std::wstring& origin) {
-  // Strip the leading dot if it exists.
-  std::wstring rewritten_origin = origin;
-  if (origin.length() >= 1 && origin[0] == '.')
-    rewritten_origin = origin.substr(1);
 
-  CookieTreeOriginNode rewritten_origin_node(rewritten_origin);
+CookieTreeOriginNode* CookieTreeRootNode::GetOrCreateOriginNode(
+    const GURL& url) {
+  CookieTreeOriginNode origin_node(url);
 
   // First see if there is an existing match.
   std::vector<CookieTreeNode*>::iterator origin_node_iterator =
       lower_bound(children().begin(),
                   children().end(),
-                  &rewritten_origin_node,
+                  &origin_node,
                   OriginNodeComparator());
 
-  if (origin_node_iterator != children().end() && rewritten_origin ==
+  if (origin_node_iterator != children().end() &&
+      CookieTreeOriginNode::TitleForUrl(url) ==
       (*origin_node_iterator)->GetTitle())
     return static_cast<CookieTreeOriginNode*>(*origin_node_iterator);
   // Node doesn't exist, create a new one and insert it into the (ordered)
   // children.
-  CookieTreeOriginNode* retval = new CookieTreeOriginNode(rewritten_origin);
+  CookieTreeOriginNode* retval = new CookieTreeOriginNode(url);
   DCHECK(model_);
   model_->Add(this, (origin_node_iterator - children().begin()), retval);
   return retval;
@@ -207,6 +200,21 @@ CookieTreeOriginNode* CookieTreeRootNode::GetOrCreateOriginNode(
 
 ///////////////////////////////////////////////////////////////////////////////
 // CookieTreeOriginNode, public:
+
+// static
+std::wstring CookieTreeOriginNode::TitleForUrl(
+    const GURL& url) {
+  return UTF8ToWide(url.SchemeIsFile() ? kFileOriginNodeName : url.host());
+}
+
+CookieTreeOriginNode::CookieTreeOriginNode(const GURL& url)
+    : CookieTreeNode(TitleForUrl(url)),
+      cookies_child_(NULL),
+      databases_child_(NULL),
+      local_storages_child_(NULL),
+      appcaches_child_(NULL),
+      url_(url) {}
+
 
 CookieTreeCookiesNode* CookieTreeOriginNode::GetOrCreateCookiesNode() {
   if (cookies_child_)
@@ -239,6 +247,19 @@ CookieTreeAppCachesNode* CookieTreeOriginNode::GetOrCreateAppCachesNode() {
   appcaches_child_ = new CookieTreeAppCachesNode;
   AddChildSortedByTitle(appcaches_child_);
   return appcaches_child_;
+}
+
+void CookieTreeOriginNode::CreateContentException(
+    HostContentSettingsMap* content_settings, ContentSetting setting) const {
+  if (CanCreateContentException()) {
+    content_settings->AddExceptionForURL(url_,
+                                         CONTENT_SETTINGS_TYPE_COOKIES,
+                                         setting);
+  }
+}
+
+bool CookieTreeOriginNode::CanCreateContentException() const {
+  return !url_.SchemeIsFile();
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -294,13 +315,13 @@ void CookieTreeNode::AddChildSortedByTitle(CookieTreeNode* new_child) {
 // CookiesTreeModel, public:
 
 CookiesTreeModel::CookiesTreeModel(
-    Profile* profile,
+    net::CookieMonster* cookie_monster,
     BrowsingDataDatabaseHelper* database_helper,
     BrowsingDataLocalStorageHelper* local_storage_helper,
     BrowsingDataAppCacheHelper* appcache_helper)
     : ALLOW_THIS_IN_INITIALIZER_LIST(TreeNodeModel<CookieTreeNode>(
           new CookieTreeRootNode(this))),
-      profile_(profile),
+      cookie_monster_(cookie_monster),
       appcache_helper_(appcache_helper),
       database_helper_(database_helper),
       local_storage_helper_(local_storage_helper),
@@ -376,19 +397,21 @@ void CookiesTreeModel::LoadCookies() {
 void CookiesTreeModel::LoadCookiesWithFilter(const std::wstring& filter) {
   // mmargh mmargh mmargh!
 
-  // Since we are running on the UI thread don't call GetURLRequestContext().
-  net::CookieMonster* cookie_monster =
-      profile_->GetRequestContext()->GetCookieStore()->GetCookieMonster();
-
-  all_cookies_ = cookie_monster->GetAllCookies();
+  all_cookies_ = cookie_monster_->GetAllCookies();
   CookieTreeRootNode* root = static_cast<CookieTreeRootNode*>(GetRoot());
   for (CookieList::iterator it = all_cookies_.begin();
        it != all_cookies_.end(); ++it) {
-    std::wstring origin_node_name = UTF8ToWide(it->first);
+    std::string origin_host = it->first;
+    if (origin_host.length() > 1 && origin_host[0] == '.')
+      origin_host = it->first.substr(1);
+    // We treat secure cookies just the same as normal ones.
+    GURL origin(std::string(chrome::kHttpScheme) +
+                chrome::kStandardSchemeSeparator + origin_host + "/");
     if (!filter.size() ||
-        (origin_node_name.find(filter) != std::wstring::npos)) {
+        (CookieTreeOriginNode::TitleForUrl(origin).find(filter) !=
+         std::string::npos)) {
       CookieTreeOriginNode* origin_node =
-          root->GetOrCreateOriginNode(origin_node_name);
+          root->GetOrCreateOriginNode(origin);
       CookieTreeCookiesNode* cookies_node =
           origin_node->GetOrCreateCookiesNode();
       CookieTreeCookieNode* new_cookie = new CookieTreeCookieNode(&*it);
@@ -409,11 +432,15 @@ void CookiesTreeModel::DeleteAllStoredObjects() {
 }
 
 void CookiesTreeModel::DeleteCookieNode(CookieTreeNode* cookie_node) {
+  if (cookie_node == GetRoot())
+    return;
   cookie_node->DeleteStoredObjects();
   // find the parent and index
   CookieTreeNode* parent_node = cookie_node->GetParent();
   int cookie_node_index = parent_node->IndexOfChild(cookie_node);
   delete Remove(parent_node, cookie_node_index);
+  if (parent_node->GetChildCount() == 0)
+    DeleteCookieNode(parent_node);
 }
 
 void CookiesTreeModel::UpdateSearchResults(const std::wstring& filter) {
@@ -465,7 +492,7 @@ void CookiesTreeModel::PopulateAppCacheInfoWithFilter(
     if (filter.empty() ||
         (origin_node_name.find(filter) != std::wstring::npos)) {
       CookieTreeOriginNode* origin_node =
-          root->GetOrCreateOriginNode(origin_node_name);
+          root->GetOrCreateOriginNode(origin->first);
       CookieTreeAppCachesNode* appcaches_node =
           origin_node->GetOrCreateAppCachesNode();
 
@@ -495,17 +522,13 @@ void CookiesTreeModel::PopulateDatabaseInfoWithFilter(
   for (DatabaseInfoList::iterator database_info = database_info_list_.begin();
        database_info != database_info_list_.end();
        ++database_info) {
-    // Determine which 'origin' node to place each 'info' in.
-    std::wstring origin_node_name;
-    if (database_info->IsFileSchemeData())
-      origin_node_name = UTF8ToWide(kFileOriginNodeName);
-    else
-      origin_node_name = UTF8ToWide(database_info->host);
+    GURL origin(database_info->origin);
 
     if (!filter.size() ||
-        (origin_node_name.find(filter) != std::wstring::npos)) {
+        (CookieTreeOriginNode::TitleForUrl(origin).find(filter) !=
+         std::wstring::npos)) {
       CookieTreeOriginNode* origin_node =
-          root->GetOrCreateOriginNode(origin_node_name);
+          root->GetOrCreateOriginNode(origin);
       CookieTreeDatabasesNode* databases_node =
           origin_node->GetOrCreateDatabasesNode();
       databases_node->AddDatabaseNode(
@@ -532,17 +555,13 @@ void CookiesTreeModel::PopulateLocalStorageInfoWithFilter(
        local_storage_info_list_.begin();
        local_storage_info != local_storage_info_list_.end();
        ++local_storage_info) {
-    // Determine which 'origin' node to place each 'info' in.
-    std::wstring origin_node_name;
-    if (local_storage_info->IsFileSchemeData())
-      origin_node_name = UTF8ToWide(kFileOriginNodeName);
-    else
-      origin_node_name = UTF8ToWide(local_storage_info->host);
+    GURL origin(local_storage_info->origin);
 
     if (!filter.size() ||
-        (origin_node_name.find(filter) != std::wstring::npos)) {
+        (CookieTreeOriginNode::TitleForUrl(origin).find(filter) !=
+         std::wstring::npos)) {
       CookieTreeOriginNode* origin_node =
-          root->GetOrCreateOriginNode(origin_node_name);
+          root->GetOrCreateOriginNode(origin);
       CookieTreeLocalStoragesNode* local_storages_node =
           origin_node->GetOrCreateLocalStoragesNode();
       local_storages_node->AddLocalStorageNode(
